+++ installer/__init__.py
@@ -1,4 +1,3 @@
-"""Deployer package."""
+"""Installer package."""
 
 __version__ = "1.2.0"
-